Project refactoring
Moved the arduino project into the src folder. The robot is now step-controllable, from within the arduino file with the global n_steps variable. This variable is going to be changed to a paramater for each direction function, so /cmd_vel can control the number of steps to be taken.