Finishes troubleshooting the lidar
Any trouble with the error  can be reduced to ROS not recognizing the leddar library. The best fix currently is to move the actual driver. This can be found (if installed with pip) in . A more permanent fix should be found. Probably defining certain paths or using rosdep, with which I've had no succes yet.