Arduino code update
Small update to arduino c ode. Fixed a bug where the speed was not controllable with cmd_vel. Twe speedfactor variable should be tweaked to get the desired speeds to comply with the cmd_vel standards. They are currently set so that the robot roughly reflects the simulation in rviz. With the slam node, this should suffice.